fix(run): keep trusted post-exec gate and unknown commit state
Close CSA review HIGHs on #3121: propagate trusted
CSA_NO_POST_EXEC_GATE after env scrub, and persist
commit_created as Option so unknown is not folded to false.

Fold the extra command-builder bool into ExecuteOptions
so clippy too-many-arguments stays at the 7-arg cap.
Extract tiny transport helpers to stay under monolith caps
without raising scripts/monolith/baseline.toml.

Closes #3121
