+++ crates/cli-sub-agent/src/run_cmd_uncommitted_require_commit.rs
@@ -60,7 +60,9 @@ pub(super) fn contract_failure_reason_with_state(
 pub(super) fn persisted_contract_failure_reason(
     recovery: &csa_session::RequireCommitRecoveryDiagnostic,
 ) -> &'static str {
-    if recovery.suggested_recovery_action == PARTIAL_COMMIT_PLUS_DIRTY_WORK {
+    if recovery.commit_created.is_none() && recovery.dirty_worktree {
+        super::REQUIRE_COMMIT_REASON
+    } else if recovery.suggested_recovery_action == PARTIAL_COMMIT_PLUS_DIRTY_WORK {
         super::REQUIRE_COMMIT_PARTIAL_REASON
     } else if recovery.suggested_recovery_action == TRACKED_DIRTY_WORK_WITHOUT_COMMIT {
         super::REQUIRE_COMMIT_DIRTY_REASON
@@ -72,7 +74,7 @@ pub(super) fn persisted_contract_failure_reason(
         == super::REQUIRE_COMMIT_SANDBOX_HOOK_RECOVERY_ACTION
     {
         super::REQUIRE_COMMIT_SANDBOX_HOOK_REASON
-    } else if recovery.dirty_worktree && recovery.commit_created {
+    } else if recovery.dirty_worktree && recovery.commit_created == Some(true) {
         super::REQUIRE_COMMIT_PARTIAL_REASON
     } else if recovery.dirty_worktree {
         super::REQUIRE_COMMIT_DIRTY_REASON
@@ -190,7 +192,7 @@ pub(super) fn build_recovery_diagnostic_for_state(
     csa_session::RequireCommitRecoveryDiagnostic {
         require_commit: true,
         sa_mode,
-        commit_created: commit_created.unwrap_or(false),
+        commit_created,
         dirty_worktree: changes.is_some(),
         changed_paths: changes
             .map(|changes| {
