+++ crates/cli-sub-agent/src/session_cmds_daemon_wait_summary_kill_tests.rs
@@ -138,7 +138,7 @@ last_accessed = "2026-06-01T00:00:00Z"
     result.require_commit_recovery = Some(csa_session::RequireCommitRecoveryDiagnostic {
         require_commit: true,
         sa_mode: Some(false),
-        commit_created: false,
+        commit_created: Some(false),
         dirty_worktree: true,
         changed_paths: vec!["crates/verbatim-daemon/src/main.rs".to_string()],
         changed_paths_truncated: 0,
