+++ crates/csa-executor/src/transport_openai_compat.rs
@@ -262,6 +262,7 @@ impl Transport for OpenaiCompatTransport {
                 thinking_budget: None,
                 subtree_pin: subtree_pin.cloned(),
                 allow_git_push,
+                no_post_exec_gate: false,
                 cancellation: None,
             },
         )
