+++ crates/csa-executor/src/transport_tests_gemini_fallback.rs
@@ -133,6 +133,7 @@ async fn test_execute_falls_back_to_api_key_after_all_retries_exhausted() {
         thinking_budget: None,
         subtree_pin: None,
         allow_git_push: false,
+        no_post_exec_gate: false,
         cancellation: None,
     };
